requirements.txt
gspread
jinja2
mailchimp3
oauth2client
py-trello
pygithub
pystache
python-dotenv
requests
slackclient
